Backend/requirements.txt
psycopg2-binary
Flask
flask-cors
paho-mqtt 